Add mp3 support in rails. Frontend implementation necessary

+++ backend/app/controllers/response_controller.rb
@@ -1,22 +1,32 @@
 class ResponseController < ApplicationController
   def text_response
     questionText = params[:question]
-    path = "#{Rails.root}/app/scripts/temp/input.txt"
-    
+
+    path = "#{Rails.root}/app/scripts/temp/input.txt"  
     File.open(path, "w") do |f|
       f.write(questionText)
     end
-    system("cd app/scripts && python3 textToText.py")
-    
-    path = "#{Rails.root}/app/scripts/temp/output.txt"
-    
-    response = "Failed!"
-    File.open(path) do |f|
-      response = f.read
-    end
-    File.delete(path)
 
-    render json: { answer: response, question: questionText }
+    if params[:speech].present? && params[:speech] == "true"
+      system("cd app/scripts && python3 textToSpeech.py")
+      path = "#{Rails.root}/app/scripts/temp/output.mp3"
+      
+      response.headers["Content-Disposition"] = 'attachment; filename="output.mp3"'
+      response.headers["Content-Type"] = "audio/mpeg"
+
+      send_file path, disposition: 'attachment'
+    else
+      response = "Failed!"
+      system("cd app/scripts && python3 textToText.py")
+      path = "#{Rails.root}/app/scripts/temp/output.txt"
+      
+      File.open(path) do |f|
+        response = f.read
+      end
+      File.delete(path)
+
+      render json: { answer: response, question: questionText }
+    end
   end
 
   def audio_response
